Relocate inode/directory mime association; use Files' context menu
Use a separate .desktop to handle inode/directory for Code--the
same way Terminal does this. Other programs should not interfere
with Files' (or user's configured file browser) ownership of
inode/directory. This causes problems, like Plank's Trash Can
opening the trash directory in the wrong program.

When right clicking a folder in Files, the "Open in..." context
menu will direct users to Code.