Expand comment about "open here" shortcut
Fully explain the need to do this (the same thing that was done for Terminal). Since other distributions and desktop environments cannot be relied on to provide overrides that ensure a specific file manager gets priority for the inode/directory association, Code must attempt to avoid making itself the default handler.