Restart shell in terminal pane if it exits (#1618)
Fixes #457

There was already code to hide the terminal if it's child exited but it had regressed and was not working.

This PR makes the terminal hide again if the shell exits.

It also restarts the shell at the last saved location and clears the screen ready for it to be used again. To enable this, the shell location is saved to settings every time it changes through app action instead of just at closing the app. (NOTE: Change of location through the shell cd command is not detected so not saved)